pyproject.toml
[project]
name = "pr-story"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.2",
    "fastapi>=0.115.12",
    "h11>=0.14.0",
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
